pyproject.toml
[tool.poetry]
name = "hackcamp2021"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pip = "^21.3.1"
bs4 = "^0.0.1"
strings-pkg = "^0.0.1"
spacy = "^3.2.0"
en-core-web-sm-mirror = "^2.2.5"
urlopen = "^1.0.0"
Flask = "^2.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
